+++ bin/migrate-oats-data/documents/__init__.py
@@ -1,4 +1,4 @@
-from .documents import *
-from .application_documents import *
-from .documents_noi import *
-from .noi_documents import *
+from .alcs_documents_to_app_documents import *
+from .oats_documents_to_alcs_documents_app import *
+from .alcs_documents_to_noi_documents import *
+from .oats_documents_to_alcs_documents_noi import *
