+++ bin/migrate-oats-data/sql/documents_app/oats_documents_to_alcs_documents_app.sql
